Extract dedicated GherkinTableRowLine
