+++ java/src/test/java/io/cucumber/gherkin/GherkinTagLineTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import java.util.Collections;
 import java.util.List;
 
-import static io.cucumber.gherkin.GherkinTagLine.parseTags;
+import static io.cucumber.gherkin.GherkinTagLine.parse;
 import static java.util.Arrays.asList;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -16,7 +16,7 @@ class GherkinTagLineTest {
 
     @Test
     void allows_any_non_space_characters_in_a_tag() {
-        List<GherkinLineSpan> gherkinLineSpans = parseTags(3, "@foo:bar  @zap\uD83E\uDD52yo", line);
+        List<GherkinLineSpan> gherkinLineSpans = parse(3, "@foo:bar  @zap\uD83E\uDD52yo", line);
 
         assertEquals(asList(
                 new GherkinLineSpan(4, "@foo:bar"),
@@ -26,7 +26,7 @@ class GherkinTagLineTest {
 
     @Test
     void finds_tags() {
-        List<GherkinLineSpan> gherkinLineSpans = parseTags(0, "@this @is @a @tag", line);
+        List<GherkinLineSpan> gherkinLineSpans = parse(0, "@this @is @a @tag", line);
 
         assertEquals(asList(
                 new GherkinLineSpan(1, "@this"),
@@ -38,29 +38,29 @@ class GherkinTagLineTest {
 
     @Test
     void throws_on_tags_with_spaces() {
-        assertThrows(ParserException.class, () -> parseTags(0, "@this @is @a space separated @tag", line));
+        assertThrows(ParserException.class, () -> parse(0, "@this @is @a space separated @tag", line));
     }
 
     @Test
     void throws_on_tags_with_leading_spaces() {
-        assertThrows(ParserException.class, () -> parseTags(0, "@ leadingSpace", line));
+        assertThrows(ParserException.class, () -> parse(0, "@ leadingSpace", line));
     }
 
     @Test
     void ignores_empty_tag() {
-        List<GherkinLineSpan> gherkinLineSpans = parseTags(0, "@", line);
+        List<GherkinLineSpan> gherkinLineSpans = parse(0, "@", line);
         assertEquals(Collections.emptyList(), gherkinLineSpans);
     }
 
     @Test
     void ignores_empty_tags() {
-        List<GherkinLineSpan> gherkinLineSpans = parseTags(0, "@@", line);
+        List<GherkinLineSpan> gherkinLineSpans = parse(0, "@@", line);
         assertEquals(Collections.emptyList(), gherkinLineSpans);
     }
 
     @Test
     void finds_tags__trim_whitespace() {
-        List<GherkinLineSpan> gherkinLineSpans = parseTags(4, "@this @is  @a @tag", line);
+        List<GherkinLineSpan> gherkinLineSpans = parse(4, "@this @is  @a @tag", line);
         assertEquals(asList(
                 new GherkinLineSpan(5, "@this"),
                 new GherkinLineSpan(11, "@is"),
@@ -71,7 +71,7 @@ class GherkinTagLineTest {
 
     @Test
     void finds_tags__comment_after_tag() {
-        List<GherkinLineSpan> gherkinLineSpans = parseTags(0, "@this @is #acomment", line);
+        List<GherkinLineSpan> gherkinLineSpans = parse(0, "@this @is #acomment", line);
         assertEquals(asList(
                 new GherkinLineSpan(1, "@this"),
                 new GherkinLineSpan(7, "@is")
@@ -80,7 +80,7 @@ class GherkinTagLineTest {
 
     @Test
     void finds_tags__comment_inside_tag_preceded_by_nbsp() {
-        List<GherkinLineSpan> gherkinLineSpans = parseTags(0, "@this @is\u202F#acomment", line);
+        List<GherkinLineSpan> gherkinLineSpans = parse(0, "@this @is\u202F#acomment", line);
         assertEquals(asList(
                 new GherkinLineSpan(1, "@this"),
                 new GherkinLineSpan(7, "@is")
@@ -89,7 +89,7 @@ class GherkinTagLineTest {
 
     @Test
     void finds_tags__comment_inside_tag() {
-        List<GherkinLineSpan> gherkinLineSpans = parseTags(0, "@comment_tag#2 #a comment", line);
+        List<GherkinLineSpan> gherkinLineSpans = parse(0, "@comment_tag#2 #a comment", line);
         assertEquals(asList(
                 new GherkinLineSpan(1, "@comment_tag#2")
         ), gherkinLineSpans);
@@ -97,7 +97,7 @@ class GherkinTagLineTest {
 
     @Test
     void finds_tags__commented_before_tag() {
-        List<GherkinLineSpan> gherkinLineSpans = parseTags(0, "@this @is #@a commented tag", line);
+        List<GherkinLineSpan> gherkinLineSpans = parse(0, "@this @is #@a commented tag", line);
         assertEquals(asList(
                 new GherkinLineSpan(1, "@this"),
                 new GherkinLineSpan(7, "@is")
@@ -106,7 +106,7 @@ class GherkinTagLineTest {
 
     @Test
     void finds_tags__commented_multiple_tags() {
-        List<GherkinLineSpan> gherkinLineSpans = parseTags(0, "@this @is #@a @commented @sequence of tags", line);
+        List<GherkinLineSpan> gherkinLineSpans = parse(0, "@this @is #@a @commented @sequence of tags", line);
         assertEquals(asList(
                 new GherkinLineSpan(1, "@this"),
                 new GherkinLineSpan(7, "@is")
@@ -115,15 +115,15 @@ class GherkinTagLineTest {
 
     @Test
     void finds_tags_with_number() {
-        List<GherkinLineSpan> gherkinLineSpans = parseTags(0, "@ISSUE#123", line);
+        List<GherkinLineSpan> gherkinLineSpans = parse(0, "@ISSUE#123", line);
         assertEquals(asList(
                 new GherkinLineSpan(1, "@ISSUE#123")
         ), gherkinLineSpans);
     }
 
     @Test
-    void parseTags_returns_empty_list_when_empty_line() {
-        List<GherkinLineSpan> gherkinLineSpans = parseTags(0, "", line);
+    void parse_returns_empty_list_when_empty_line() {
+        List<GherkinLineSpan> gherkinLineSpans = parse(0, "", line);
         assertEquals(Collections.emptyList(), gherkinLineSpans);
     }
 }
